rename: layout, button 폴더명 변경
