+++ src/app/(route)/(home)/layout.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren } from 'react';
 
-import { FooterLayout } from '@/components/layout/Footer';
+import { FooterLayout } from '@/components/Layout/Footer';
 
 export default function HomeLayout({ children }: PropsWithChildren) {
   return <FooterLayout>{children}</FooterLayout>;
